// temp-job-frontend/features/education-level/schemas/education-level.schema.ts
import { z } from "zod";

export const EducationLevelSchema = z.object({
  id: z.number({ message: "กรุณาเลือกวุฒิการศึกษา" }),
  created_at: z.string(),
  title: z.string(),
  power: z.number(),
});

export type EducationLevel = z.infer<typeof EducationLevelSchema>;